处理了光照传感器的驱动链路中I2C可能发生的错误

+++ BSP/Applayer/app_algorithm.c
@@ -1,23 +1,23 @@
 
 
 #include "app_algorithm.h"
-
+#include "app_manager.h"
 
 /**
  * @brief 映射光照强度到灯带亮度 低通滤波
  */
 void remap_lux_to_brightness(void)
 {
-    uint8_t brightness = system.system_data.rgb_data.brightness;
+    float brightness = (float)system.system_data.rgb_data.brightness;
     float lux = system.system_data.ls_data.lux;
     float brightness_filter = system.system_data.rgb_data.filter_lux;
 
     float target_brightness = RGB_MAX_BRIGHTNESS * sqrt(lux / LS_LUX_MAX); // 非线性映射，亮度变化更明显
 
-    brightness = (uint8_t)(brightness_filter * brightness + \
-    (1-brightness_filter)*target_brightness);
+    brightness = brightness_filter * brightness + \
+    (1.0f - brightness_filter)*target_brightness;
 
-    system.system_data.rgb_data.brightness = brightness;
+    system.system_data.rgb_data.brightness = (uint8_t)brightness;
 }
